chore: move new api related things to new v2 files
