+++ lms/djangoapps/instructor/views/api_urls.py
@@ -33,7 +33,7 @@ v2_api_urls = [
     ),
     re_path(
       rf'^courses/{COURSE_ID_PATTERN}/change_due_date$',
-      api.ChangeDueDateV2.as_view(),
+      api_v2.ChangeDueDateV2.as_view(),
       name='change_due_date_v2'
     ),
 ]
