+++ lms/djangoapps/instructor/views/api_v2.py
@@ -14,17 +14,30 @@ from rest_framework import status
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
+from django.utils.decorators import method_decorator
+from django.views.decorators.cache import cache_control
+from django.views.decorators.csrf import ensure_csrf_cookie
+from django.utils.html import strip_tags
+from django.utils.translation import gettext as _
+from common.djangoapps.util.json_request import JsonResponse, JsonResponseBadRequest
 
 from lms.djangoapps.courseware.tabs import get_course_tab_list
 from lms.djangoapps.instructor import permissions
-from lms.djangoapps.instructor.views.api import get_student_from_identifier
+from lms.djangoapps.instructor.views.api import _display_unit, get_student_from_identifier
 from lms.djangoapps.instructor.views.instructor_task_helpers import extract_task_features
 from lms.djangoapps.instructor_task import api as task_api
 from openedx.core.lib.api.view_utils import DeveloperErrorViewMixin
 from openedx.core.lib.courses import get_course_by_id
 from .serializers_v2 import (
     InstructorTaskListSerializer,
-    CourseInformationSerializer
+    CourseInformationSerializer,
+    BlockDueDateSerializerV2,
+)
+from .tools import (
+    DashboardError,
+    find_unit,
+    parse_datetime,
+    set_due_date_extension,
 )
 
 log = logging.getLogger(__name__)
@@ -269,3 +282,56 @@ class InstructorTaskListView(DeveloperErrorViewMixin, APIView):
         tasks_data = [extract_task_features(task) for task in tasks]
         serializer = InstructorTaskListSerializer({'tasks': tasks_data})
         return Response(serializer.data, status=status.HTTP_200_OK)
+
+@method_decorator(cache_control(no_cache=True, no_store=True, must_revalidate=True), name='dispatch')
+class ChangeDueDateV2(APIView):
+    """
+    Grants a due date extension to a student for a particular unit.
+    this version works with a new payload that is JSON and more up to date.
+    """
+    permission_classes = (IsAuthenticated, permissions.InstructorPermission)
+    permission_name = permissions.GIVE_STUDENT_EXTENSION
+    serializer_class = BlockDueDateSerializerV2
+
+    @method_decorator(ensure_csrf_cookie)
+    def post(self, request, course_id):
+        """
+        Grants a due date extension to a learner for a particular unit.
+
+        params:
+            blockId (str): The URL related to the block that needs the due date update.
+            due_datetime (str): The new due date and time for the block.
+            email_or_username (str): The email or username of the learner whose access is being modified.
+        """
+        serializer_data = self.serializer_class(data=request.data)
+        if not serializer_data.is_valid():
+            return JsonResponseBadRequest({'error': _('All fields must be filled out')})
+
+        learner = serializer_data.validated_data.get('email_or_username')
+        if not learner:
+            response_payload = {
+                'error': _(
+                    'Could not find learner matching identifier: {email_or_username}'
+                ).format(email_or_username=request.data.get("email_or_username"))
+            }
+            return JsonResponse(response_payload, status=status.HTTP_404_NOT_FOUND)
+
+        due_datetime = serializer_data.validated_data.get('due_datetime')
+        try:
+            due_date = parse_datetime(due_datetime)
+        except DashboardError:
+            return JsonResponseBadRequest({'error': _('The extension due date and time format is incorrect')})
+
+        course = get_course_by_id(CourseKey.from_string(course_id))
+        unit = find_unit(course, serializer_data.validated_data.get('block_id'))
+        reason = strip_tags(serializer_data.validated_data.get('reason', ''))
+        try:
+            set_due_date_extension(course, unit, learner, due_date, request.user, reason=reason)
+        except Exception as error:  # pylint: disable=broad-except
+            return JsonResponseBadRequest({'error': str(error)})
+
+        return JsonResponse({
+            'message': _(
+            'Successfully changed due date for learner {0} for {1} '
+            'to {2}').format(learner.profile.name, _display_unit(unit), due_date.strftime('%Y-%m-%d %H:%M'))
+        })
