+++ lms/djangoapps/instructor/views/serializer.py
@@ -335,43 +335,6 @@ class BlockDueDateSerializer(serializers.Serializer):
         if disable_due_datetime:
             self.fields['due_datetime'].required = False
 
-class BlockDueDateSerializerV2(serializers.Serializer):
-    """
-    Serializer for handling block due date updates for a specific student.
-    Fields:
-        block_id (str): The ID related to the block that needs the due date update.
-        due_datetime (str): The new due date and time for the block.
-        email_or_username (str): The email or username of the student whose access is being modified.
-        reason (str): Reason why updating this.
-    """
-    block_id = serializers.CharField()
-    due_datetime = serializers.CharField()
-    email_or_username = serializers.CharField(
-        max_length=255,
-        help_text="Email or username of user to change access"
-    )
-    reason = serializers.CharField(required=False)
-
-    def validate_email_or_username(self, value):
-        """
-        Validate that the email_or_username corresponds to an existing user.
-        """
-        try:
-            user = get_student_from_identifier(value)
-        except User.DoesNotExist:
-            return None
-
-        return user
-
-    def __init__(self, *args, **kwargs):
-        # Get context to check if `due_datetime` should be optional
-        disable_due_datetime = kwargs.get('context', {}).get('disable_due_datetime', False)
-        super().__init__(*args, **kwargs)
-        if disable_due_datetime:
-            self.fields['due_datetime'].required = False
-
-
-
 class ProblemResetSerializer(UniqueStudentIdentifierSerializer):
     """
     serializer for resetting problem.
